modify some code & organize in some folder

+++ hw1/kernel/multiply.c
@@ -1,7 +1,7 @@
 #include <linux/kernel.h>
 #include <linux/linkage.h>
 
-asmlinkage int sys_multiply(long a, long b)
+asmlinkage long sys_multiply(long a, long b)
 {
 	return a * b;
 }
